# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "JSON-only operator surface of the arXMCP server: /api/v1 notebook management aliases, health/readiness/status probes, and the /bridge/contracts handshake. Generated OFFLINE by tools/dump_openapi.py (IF-1); the runtime openapi_url is disabled by design (Threat 4).",
    "title": "arXMCP operator API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events/stream": {
      "get": {
        "description": "The single multiplexed live stream (gap R8; AC-A.14; IF-2).\n\n``topics`` is a comma-separated subset of\n``logs,requests,sessions,ingest``; unknown names are a 422 (a\ntypo'd topic silently delivering nothing would be a debugging\ntrap). Per-client queue cap + drop-oldest are enforced in\n:class:`server.observability.events.EventBus` \u2014 a stalled\nconsumer's server-side footprint is bounded by the queue cap.\n\nThe response is exempt from ``BodySizeCapMiddleware`` (the stream\nis unbounded by design \u2014 see the segment-exact carve-out in\n``server/main.py::_is_exempt_path``; flagged for the issue-#9\nscope amendment).",
        "operationId": "events_stream_api_v1_events_stream_get",
        "parameters": [
          {
            "in": "query",
            "name": "topics",
            "required": false,
            "schema": {
              "default": "logs,requests,sessions,ingest",
              "title": "Topics",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events Stream",
        "tags": [
          "observability"
        ]
      }
    },
    "/api/v1/ingest-events": {
      "get": {
        "description": "Additive ingest/parse stage events (gap R4; AC-A.15). The\nlegacy tri-state poll surfaces are untouched \u2014 this ring is a new,\nparallel read model.",
        "operationId": "list_ingest_events_api_v1_ingest_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "since_seq",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since Seq"
            }
          },
          {
            "in": "query",
            "name": "slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Ingest Events Api V1 Ingest Events Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Ingest Events",
        "tags": [
          "observability"
        ]
      }
    },
    "/api/v1/logs/tail": {
      "get": {
        "description": "Snapshot of the in-process log ring (gap R2; AC-A.12).\n\nEvery record here passed ``RedactionFilter`` BEFORE entering the\nring (handler-level filter \u2014 the ordering contract in\n``logging_setup``). ``level`` filters case-insensitively on the\nrecord's level name (a facet, not a query language).",
        "operationId": "logs_tail_api_v1_logs_tail_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "since_seq",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since Seq"
            }
          },
          {
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Logs Tail Api V1 Logs Tail Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logs Tail",
        "tags": [
          "observability"
        ]
      }
    },
...
    "/api/v1/requests": {
      "get": {
        "description": "Last-N request events, newest first (gap R1; AC-A.11).\n\n``since_seq`` returns only events strictly newer than the cursor \u2014\nthe polling degraded mode's gap-free catch-up (within ring\ncapacity; a seq jump larger than the ring means events aged out).",
        "operationId": "list_request_events_api_v1_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "since_seq",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since Seq"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Request Events Api V1 Requests Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Request Events",
        "tags": [
          "observability"
        ]
      }
    },
    "/api/v1/sessions": {
      "get": {
        "description": "Read-only projection of the MCP session registry with per-cap\nremaining counts (gaps R3/R7; AC-A.13). The registry itself stays\nmodule-private in :mod:`server.session` \u2014 this endpoint serializes\na point-in-time copy and never mutates state.",
        "operationId": "list_sessions_api_v1_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Sessions Api V1 Sessions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "observability"
        ]
      }
    },
    "/api/v1/tool-calls": {
      "get": {
        "description": "Newest-first page over the append-only ``tool_calls`` audit\nstore (AC-A.9's read side). 503 while the store is not yet bound\n(startup) \u2014 a probe-friendly signal, not an error state.",
        "operationId": "list_tool_calls_api_v1_tool_calls_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Tool Calls Api V1 Tool Calls Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tool Calls",
        "tags": [
          "observability"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:7733"
    }
  ]
}